Create Find Most Expensive Stored Procedure Sql Server
Problem: An important step in optimizing a system is to take a holistic approach and look at stored procedures that are called very often. These procedures can often be the backbone of a system. Sometimes optimizing these procedures can have a dramatic effect on relieving the physical bottlenecks of the system and improving end-user experience.

Solution: The following DMV query shows the execution count of each stored procedure, sorted by the most executed procedures first.